Restructure MDX component files
