--- src/components/layout/mdx/index.ts
@@ -1,2 +1,0 @@
-export * from './mdx';
-export * from './mdx-blockquote';
